Time: 4 ms (44.03%), Space: 6.1 MB (30.50%) - LeetHub

+++ 1359-count-all-valid-pickup-and-delivery-options/1359-count-all-valid-pickup-and-delivery-options.cpp
@@ -14,6 +14,6 @@ public:
             dp[i] = (dp[i-1]* sum)%mod;
         }
         
-        return (int) dp[n];
+        return dp[n];
     }
 };
